cotisa/requirements.txt
Django==4.2.7
mysqlclient==2.2.0
Pillow==10.1.0
python-decouple==3.8
django-cors-headers==4.3.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
requests==2.31.0
